♻️(frontend) split InputDisplayNameOverlay component
To get rid of a context provider useless in a certain case,
InputDisplayNameOverlay has been split in two components.
