+++ src/frontend/components/Chat/SharedChatComponents/InputDisplayName/InputDisplayNameIncorrectAlert/__image_snapshots__/index-spec-tsx-input-display-name-incorrect-alert-displays-the-component-and-compares-it-with-previous-render-screenshot-1-snap.png
