+++ src/frontend/components/Chat/SharedChatComponents/InputDisplayName/InputDisplayNameIncorrectAlert/index.spec.tsx
@@ -3,6 +3,7 @@ import { render, screen } from '@testing-library/react';
 
 import { renderImageSnapshot } from 'utils/tests/imageSnapshot';
 import { wrapInIntlProvider } from 'utils/tests/intl';
+
 import { InputDisplayNameIncorrectAlert } from '.';
 
 describe('<InputDisplayNameIncorrectAlert />', () => {
